remove zod-prisma dep
