+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@auth/core": "^0.15.0",
     "@hookform/resolvers": "^3.3.1",
+    "@prisma/client": "^5.3.1",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-alert-dialog": "^1.0.4",
     "@radix-ui/react-avatar": "^1.0.3",
@@ -60,7 +61,6 @@
   "devDependencies": {
     "dotenv": "^16.3.1",
     "prisma": "^5.3.1",
-    "tsx": "^3.12.10",
-    "zod-prisma": "^0.5.4"
+    "tsx": "^3.12.10"
   }
 }
