--- prisma/zod/index.ts
@@ -1,7 +1,0 @@
-export * from "./computer"
-export * from "./account"
-export * from "./session"
-export * from "./user"
-export * from "./verificationtoken"
-export * from "./bot"
-export * from "./command"
