--- prisma/zod/user.ts
@@ -1,29 +1,0 @@
-import * as z from "zod"
-import { CompleteAccount, relatedAccountSchema, CompleteSession, relatedSessionSchema, CompleteBot, relatedBotSchema } from "./index"
-
-export const userSchema = z.object({
-  id: z.string(),
-  username: z.string(),
-  name: z.string().nullish(),
-  displayName: z.string().nullish(),
-  email: z.string().nullish(),
-  emailVerified: z.date().nullish(),
-  image: z.string().nullish(),
-})
-
-export interface CompleteUser extends z.infer<typeof userSchema> {
-  accounts: CompleteAccount[]
-  sessions: CompleteSession[]
-  bot: CompleteBot[]
-}
-
-/**
- * relatedUserSchema contains all relations on your model in addition to the scalars
- *
- * NOTE: Lazy required in case of potential circular dependencies within schema
- */
-export const relatedUserSchema: z.ZodSchema<CompleteUser> = z.lazy(() => userSchema.extend({
-  accounts: relatedAccountSchema.array(),
-  sessions: relatedSessionSchema.array(),
-  bot: relatedBotSchema.array(),
-}))
