--- prisma/zod/verificationtoken.ts
@@ -1,7 +1,0 @@
-import * as z from "zod"
-
-export const verificationTokenSchema = z.object({
-  identifier: z.string(),
-  token: z.string(),
-  expires: z.date(),
-})
